Almost all works.
As mentioned previously I had to shuffle it to get the arrangement I
wanted for landscape, I will have to fix that for next quarter. Also I
have strange issues with on of the pages blocking its opposite from
printing. I'll probably just scrap the bitcoin page and make it a new to
avoid this issue. If it does come back note that if you print first to
post script it will print fine.