PDF created!
I have deleted some of the other code necessary for the creation of the
pdfs that will later be compiled, but I have succesfully compiled all my
pages from last quarter into one pdf, with correct collation. Bam!
Huzzah! The other part of this, creating the pdfs from the HTML, is
relatively trivial.

Next quarter I might even refine this a little further so that the
creation of the HTML is a little less hacky. It would be pretty neat to
get this to the point where it is something that could be shared. Allow
for other people to create their own bullet journals with the templates.