Done for now.
I've got it all working for now. The pages all compile and output to
pdfs which are then consumed and compiled into one big pdf that can then
be printed. This leaves me in a great condition for when the next
quarter comes and I need to make the next batch of pages.