Used ChatGPT to make more templates.
I thing this will be a very fruitful relationship. The templates are
the part of this I dread the most so I avoided making new ones all
the time if I could help it. ChatGPT not only made it faster for me
but it used the css grid which made item_grid parametric.