I have abstracted graph a bit.
I've got my advent of code page looking great but in the process I
have ruined the running page. So I'm going to go back and rework
them both a bit. I think I need to make more templates that
eventually result in a graph as there are a lot of basic ideas
here. Running is a days of week graph that spans the quarter.
Advent of Code is a graph that spans a month. These are clear
reusable concepts that can make the yaml much less complicated.